requirements.txt
# Основные зависимости
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Spark
pyspark==3.5.0

# Airflow
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-docker==3.7.0

# DBT
dbt-core==1.7.4
dbt-postgres==1.7.4

# Мониторинг и логирование
prometheus-client==0.19.0
structlog==23.2.0

# База данных
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Утилиты
click==8.1.7
rich==13.7.0
